# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IntentFence Policy API",
    "version": "0.13.0",
    "description": "Fail-closed, action-bound authorization for autonomous AI agents, MCP metadata risk scanning, and x402 payment safety with ES256-signed receipts.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "contact": { "url": "https://github.com/razel369/intentfence" },
    "x-guidance": "Immediately before a consequential MCP, HTTP, A2A, or payment action, call /api/actions/authorize and execute only if the five-minute receipt verifies and its action digest still matches. Use /api/agent-risk/scan for a free metadata-only MCP tool scan."
  },
  "servers": [
    { "url": "https://agentpass-protocol.rmalka06.chatgpt.site" }
  ],
  "security": [],
...
  "paths": {
...
    "/api/x402-readiness": {
      "post": {
        "operationId": "checkLiveX402Readiness",
        "summary": "Check a live x402 endpoint without paying it",
        "description": "Makes one bounded credential-free request to a public HTTPS endpoint, blocks redirects and private-network targets, validates the live PAYMENT-REQUIRED challenge against a price ceiling and optional payee allowlist, never pays the target, and returns a signed five-minute receipt. Costs 0.002 USDC through x402.",
        "tags": ["x402", "endpoint-readiness", "payment-safety", "security"],
        "security": [],
        "x-payment-info": { "price": { "mode": "fixed", "currency": "USD", "amount": "0.002" }, "protocols": [{ "x402": { "version": 2, "network": "eip155:8453", "asset": "USDC", "payTo": "0x833ca7dcdb6a681ddc0c15982ef0d609bceb3a5e" } }] },
        "x-x402-price": "$0.002",
        "parameters": [{ "name": "PAYMENT-SIGNATURE", "in": "header", "required": false, "description": "Base64-encoded x402 v2 payment payload. Omit on the first call to receive PAYMENT-REQUIRED.", "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/X402ReadinessRequest" }, "example": { "target_url": "https://merchant.example/paid", "method": "GET", "max_price_usdc": "0.01", "allowed_payees": ["0x1111111111111111111111111111111111111111"] } } } },
        "responses": {
          "200": { "description": "Live readiness result returned and the IntentFence fee settled", "headers": { "PAYMENT-RESPONSE": { "description": "Base64-encoded x402 settlement result.", "schema": { "type": "string" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/X402ReadinessDecision" } } } },
          "400": { "$ref": "#/components/responses/Problem" },
          "402": { "description": "Payment is required or could not be verified" },
          "409": { "$ref": "#/components/responses/Problem" },
          "413": { "$ref": "#/components/responses/Problem" },
          "415": { "$ref": "#/components/responses/Problem" },
          "500": { "$ref": "#/components/responses/Problem" },
          "503": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "Problem": {
        "description": "Problem Details error response",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
...
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "title", "status", "detail", "instance"],
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string" },
          "status": { "type": "integer", "minimum": 400, "maximum": 599 },
          "detail": { "type": "string" },
          "instance": { "type": "string" }
        }
      },
...
      "X402ReadinessRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["target_url"],
        "properties": {
          "subject": { "type": "string", "minLength": 1, "maxLength": 200 },
          "target_url": { "type": "string", "format": "uri", "maxLength": 2048, "description": "Public HTTPS endpoint on the standard port." },
          "method": { "type": "string", "enum": ["GET", "HEAD", "POST"], "default": "GET" },
          "body": {},
          "max_price_usdc": { "type": "string", "pattern": "^(?:0|[1-9][0-9]{0,11})(?:\\.[0-9]{1,6})?$", "default": "1.00" },
          "allowed_payees": { "type": "array", "minItems": 1, "maxItems": 20, "uniqueItems": true, "items": { "type": "string", "pattern": "^0x[0-9a-fA-F]{40}$" } }
        }
      },
      "X402ReadinessDecision": {
        "type": "object",
        "required": ["intentfence", "request_id", "status", "checked_at", "target", "observed", "checks", "assessment", "verification_tier", "receipt"],
        "properties": {
          "intentfence": { "type": "string", "const": "0.9" },
          "request_id": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["ready", "ready_with_review", "not_ready"] },
          "checked_at": { "type": "string", "format": "date-time" },
          "target": { "type": "object" },
          "observed": { "type": "object" },
          "checks": { "type": "array", "items": { "type": "object" } },
          "assessment": { "type": ["object", "null"] },
          "verification_tier": { "type": "string", "const": "live-x402-readiness+x402-settled" },
          "receipt": { "type": "object" }
        }
      },
...
    }
  }
}